+++ src/extension.ts
@@ -291,7 +291,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     vscode.commands.registerCommand("codemap.sort_location", () => sort(SortDirection.ByLocation));
     vscode.commands.registerCommand("codemap.sort_asc", () => sort(SortDirection.Asc));
-    vscode.commands.registerCommand("codemap.sort_desc", () => sort(SortDirection.Decs));
+    vscode.commands.registerCommand("codemap.sort_desc", () => sort(SortDirection.Desc));
 
     vscode.commands.registerCommand("codemap.mappers", () => {
         let mappers = vscode.workspace.getConfiguration("codemap");